made updates for Selenium 4.0
