+++ jmaqs-selenium/src/main/java/com/magenic/jmaqs/selenium/SeleniumUtilities.java
@@ -4,6 +4,7 @@
 
 package com.magenic.jmaqs.selenium;
 
+import com.magenic.jmaqs.base.exceptions.JMAQSRuntimeException;
 import com.magenic.jmaqs.utilities.helper.StringProcessor;
 import com.magenic.jmaqs.utilities.logging.FileLogger;
 import com.magenic.jmaqs.utilities.logging.LoggingConfig;
@@ -64,6 +65,8 @@ public class SeleniumUtilities {
         return false;
       }
 
+      testObject.getLogger().logMessage(MessageType.VERBOSE, "Before screenshot capture");
+
       // Calculate the file name with Date Time Stamp
       String directory = ((FileLogger) testObject.getLogger()).getDirectory();
       String fileNameWithoutExtension = StringProcessor
@@ -73,6 +76,7 @@ public class SeleniumUtilities {
       captureScreenshot(webDriver, testObject, directory, fileNameWithoutExtension);
 
       testObject.getLogger().logMessage(MessageType.INFORMATION, "Screenshot saved.");
+      testObject.getLogger().logMessage(MessageType.VERBOSE, "After screenshot capture");
       return true;
     } catch (Exception exception) {
       testObject.getLogger().logMessage(MessageType.ERROR,
@@ -243,6 +247,22 @@ public class SeleniumUtilities {
       webDriver.close();
     } finally {
       webDriver.quit();
+    }
+  }
+
+  /**
+   * Switches to a different window.
+   * @param testObject the test object to be used
+   * @param windowName the name of the window being switched to
+   */
+  public static void switchToWindow(SeleniumTestObject testObject, String windowName) {
+    testObject.getLogger().logMessage(MessageType.VERBOSE, "Before switching to window: %s", windowName);
+
+    try {
+      testObject.getWebDriver().switchTo().window(windowName);
+      testObject.getLogger().logMessage(MessageType.VERBOSE, "After switching to window: %s", windowName);
+    } catch (Exception e) {
+      throw new JMAQSRuntimeException(e.getMessage(), e);
     }
   }
 }
