+++ jmaqs-selenium/src/test/java/com/magenic/jmaqs/selenium/SeleniumUtilitiesUnitTest.java
@@ -23,6 +23,7 @@ import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.support.events.EventFiringDecorator;
 import org.openqa.selenium.support.events.WebDriverListener;
 import org.testng.Assert;
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 /**
@@ -31,13 +32,23 @@ import org.testng.annotations.Test;
 public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
 
         /**
+         * The web driver used in the test.
+         */
+        private WebDriver webDriver;
+
+        /**
+         * Sets up the web driver for the unit tests.
+         */
+        @BeforeMethod
+        public void setUp() {
+                webDriver = WebDriverFactory.getDefaultBrowser();
+        }
+
+        /**
          * Test capture screenshot no append.
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testCaptureScreenshotNoAppend() {
-
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -64,8 +75,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testCaptureScreenshotAppend() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -97,8 +106,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testCaptureScreenshotConsoleLogger() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         ConsoleLogger consoleLogger = new ConsoleLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, consoleLogger,
@@ -121,8 +128,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testCaptureScreenshotCustomDirectoryFileName() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -153,8 +158,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testSavePageSourceNoAppend() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -181,8 +184,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testSavePageSourceAppend() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -214,8 +215,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testSavePageSourceCustomDirectoryFileName() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         FileLogger fileLogger = (FileLogger) this.getTestObject().getLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, fileLogger,
@@ -251,8 +250,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testSavePageSourceConsoleLogger() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         ConsoleLogger consoleLogger = new ConsoleLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, consoleLogger,
@@ -279,10 +276,8 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testWebElementToWebDriver() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
-                        WebDriverListener listener = new EventHandler2(this.getLogger());
+                        WebDriverListener listener = new EventHandler(this.getLogger());
                         webDriver = new EventFiringDecorator(listener).decorate(webDriver);
 
                         ConsoleLogger consoleLogger = new ConsoleLogger();
@@ -306,8 +301,6 @@ public class SeleniumUtilitiesUnitTest extends BaseGenericTest {
          */
         @Test(groups = TestCategories.SELENIUM)
         public void testKillDriver() {
-                WebDriver webDriver = WebDriverFactory.getDefaultBrowser();
-
                 try {
                         ConsoleLogger consoleLogger = new ConsoleLogger();
                         SeleniumTestObject testObject = new SeleniumTestObject(webDriver, consoleLogger,
